fix warnings

+++ KidsMP3Player/DFMiniMp3.h
@@ -307,7 +307,7 @@ private:
 
     void sendPacket(uint8_t command, uint16_t arg = 0, uint16_t sendSpaceNeeded = c_msSendSpace)
     {
-        uint8_t out[DfMp3_Packet_SIZE] = { 0x7E, 0xFF, 06, command, 00, (arg >> 8), (arg & 0x00ff), 00, 00, 0xEF };
+        uint8_t out[DfMp3_Packet_SIZE] = { 0x7E, 0xFF, 06, command, 00, (uint8_t)(arg >> 8), (uint8_t)(arg & 0x00ff), 00, 00, 0xEF };
 
         setChecksum(out);
 
@@ -326,7 +326,7 @@ private:
         _lastSend = millis();
     }
 
-    bool readPacket(uint8_t* command, int16_t* argument)
+    bool readPacket(uint8_t* command, uint16_t* argument)
     {
         uint8_t in[DfMp3_Packet_SIZE] = { 0 };
         uint8_t read;
